--- a/src/controllers/version.rs
+++ b/src/controllers/version.rs
@@ -6,28 +6,27 @@ pub mod yank;
 use super::prelude::*;
 
 use crate::db::DieselPooledConn;
-use crate::models::{Crate, CrateVersions, Version};
-use crate::schema::versions;
+use crate::models::{Crate, Version};
 
 fn version_and_crate(req: &dyn Request) -> AppResult<(DieselPooledConn<'_>, Version, Crate)> {
-    let crate_name = &req.params()["crate_id"];
+    let crate_name = extract_crate_name(req);
+    let semver = extract_semver(req)?;
+
+    let conn = req.db_conn()?;
+    let krate = Crate::by_name(crate_name).first::<Crate>(&*conn)?;
+    let version = krate.find_version(&conn, semver)?;
+
+    Ok((conn, version, krate))
+}
+
+fn extract_crate_name(req: &dyn Request) -> &str {
+    &req.params()["crate_id"]
+}
+
+fn extract_semver(req: &dyn Request) -> AppResult<&str> {
     let semver = &req.params()["version"];
     if semver::Version::parse(semver).is_err() {
         return Err(cargo_err(&format_args!("invalid semver: {}", semver)));
     };
-
-    let conn = req.db_conn()?;
-    let krate = Crate::by_name(crate_name).first::<Crate>(&*conn)?;
-    let version = krate
-        .all_versions()
-        .filter(versions::num.eq(semver))
-        .first(&*conn)
-        .map_err(|_| {
-            cargo_err(&format_args!(
-                "crate `{}` does not have a version `{}`",
-                crate_name, semver
-            ))
-        })?;
-
-    Ok((conn, version, krate))
+    Ok(semver)
 }