@jtgeibel 2020-01-09
@sgrif If you don't have time to review the PR itself I can reassign, but I'd love to hear your opinion on this change from an architectural standpoint.  This PR focuses on endpoints with a lot of traffic or that are known to be expensive.

We could switch more endpoints over time, but even long-term I think we would avoid converting any endpoints related to authentication.  My biggest concern would be that if the follower database falls behind, then we wouldn't want to block a user from logging in (to a newly created account) or from using a newly added API token.  (Most endpoint using authentication will need write access for other reasons anyway.)  For the endpoints listed above, I don't think there are any issues if we temporarily serve some stale data.

Open questions/concerns:

* Should we add our own alerting here to monitor that the follower stays up-to-date?  I assume Heroku already has something much more robust than what we could implement, but it's worth considering.
* This does move our production environment a bit further away from our staging, CI, and development environments.
* Is this a good change, but just "too soon"?  I think that would be a very reasonable argument.  At steady-state we don't appear to be at capacity with our database, but we have been seeing the occasional spikes (a few times a day typically) where our response times jump.  In particular, I've added a "Slow Downloads (>1s)" query to Papertrail and the 48 hour velocity graph gives a good overview of the frequency and quantity of affected requests.  Given that the spikes seem to affect both web dynos at the same time and are extremely transient leads me to believe that this is not a noisy neighbor situation.  (In some cases, it is very obvious that we were hit with a lot of expensive reverse-dependency queries.  Ruling those out, the smaller spikes don't seem to align with any obvious excessive load or expensive queries that would be expected to impact other request.)