@jtgeibel 2020-01-18
> Have we confirmed that materialized views being rebuilt is replicating correctly?

I've compared `select * from recent_crate_downloads;` between the leader and follower databases and the first few pages of results were the same.   I've also tested on staging and see no issues.

I plan to deploy to production in the next hour or so.  (I recently deployed master, so we can quickly roll this back if needed.)

@bors: r=sgrif