@jtgeibel 2020-01-18
> Test timed out

It looks like travis received the job, but didn't start running any jobs until 4 minutes ago.  I've canceled those jobs, lets try again.

@bors retry