@jtgeibel 2020-01-18
Travis again waited 25+ minutes to queue the jobs, so of course they finished 2 minutes after the timeout!  There are no status issues reported for Travis, one last try for now.

@bors retry