@jtgeibel 2020-01-23
To follow up on this PR, I'd like to document some of the performance improvements I'm seeing in the Heroku PG diagnostics tab.  These numbers were collected on Monday evening and are a rough representation of the data shown there.  Since the before and after query performance data are from 2 different tabs, the I manually scrolled through the graphs to find the min and max values.  I can't guarantee that I've found the min/max values over this time range, but I'm confident the numbers below are close.

I've also dropped (and noted) up to 2 outliers from each graph, to avoid including points that were far outside the normal variance.  I have notes locally on the specific queries, but for the summary below I've elided the detailed and just provide a rough name for each query.

## Crates ordered by `downloads` descending

* Before (when run on primary)
  * 134ms to 168ms (outlier: 191ms)
* After (when run on read-only replica)
  * 125ms to 141ms (outlier: 149ms)

## Crates ordered by `updated_at` descending

* Before
  * 131ms to 449ms (outlier: 568ms)
* After
  * 126ms to 136ms (outlier: 141ms)

## Crates ordered by `name` ascending

* Before
  * 144ms to 789ms
* After
  * 129ms to 150ms

## Crates ordered by `name` ascending (with offset)

* Before
  * 142ms to 223ms (outliers: 453ms, 668ms)
* After
  * 123ms to 141ms (outlier: 312ms)

## Summary

Overall these numbers look really promising.  In many cases, the longest times when run against the replica align pretty closely with the shortest time observed when running the query against the primary database.

Additionally, I haven't received any emails from Papertrail triggered by slow (>1s) downloads since deploying this over the weekend.  Previously, it was common to receive 1 or 2 emails a day showing a small batch of slow downloads.  (That alert triggers with a threshold of 25 slow downloads in a 10 minute window.  I see only 21 slow downloads in total since the deploy, and I have an idea on what is causing that which I'll describe in a separate issue.)  

Finally, the following search can be run against the logs showing that (unexpectedly) slow queries have been virtually eliminated (as logged by PG): `program:app/postgres duration -CURRENT_TIMESTAMP -COPY -refresh_recent_crate -!=`.  (This filters our the 4 known slow queries, but its possible this accidentally filters out extra queries.)  There is only 1 batch of 6 slow queries occurring after the read-only replica deploy.